move summary sections into plan-summary folder
