+++ src/interface/src/app/plan/plan-summary/saved-scenarios/saved-scenarios.component.ts
