+++ src/interface/src/app/plan/plan-summary/summary-panel/summary-panel.component.html
